Plot.SignalXY()
